+++ libs/rust/sdk/src/actions/impls/mod.rs
@@ -6,19 +6,26 @@ use slog::debug;
 use crate::actions::Action;
 use crate::actions::ActionHook;
 use crate::AgentContext;
+use crate::Result;
 
 #[cfg(any(debug_assertions, test))]
 pub(crate) mod debug;
 mod service;
+mod shell;
 mod test;
 
 /// Register standard agent actions.
-pub fn register_std_actions(context: &AgentContext, hooks: HashMap<ActionHook, Arc<dyn Action>>) {
+pub fn register_std_actions(
+    context: &AgentContext,
+    hooks: HashMap<ActionHook, Arc<dyn Action>>,
+) -> Result<()> {
     debug!(context.logger, "Registering standard actions");
     let graceful = hooks.get(&ActionHook::StoreGracefulStop).cloned();
     self::service::register(context, graceful);
+    self::shell::register(context)?;
     self::test::register(context);
 
     #[cfg(any(debug_assertions, test))]
     self::debug::register_debug_actions(context);
+    Ok(())
 }